requirements.txt
discord.py==1.7.3
subprocess.run==0.0.8
asyncio==3.4.3